Move icon into icon subdirectory
